% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC.drcOrdinal.R
\name{AIC.drcOrdinal}
\alias{AIC.drcOrdinal}
\title{AIC Method for drcOrdinal Objects}
\usage{
\method{AIC}{drcOrdinal}(object, ..., k = 2)
}
\arguments{
\item{object}{A drcOrdinal model object}

\item{...}{Additional arguments (not used)}

\item{k}{The penalty per parameter to be used; default is 2}
}
\value{
The AIC value
}
\description{
Calculates the Akaike Information Criterion for drcOrdinal model objects.
}
