% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.drcHetVar.R
\name{BIC.drcHetVar}
\alias{BIC.drcHetVar}
\title{BIC Method for drcHetVar Objects}
\usage{
\method{BIC}{drcHetVar}(object, ...)
}
\arguments{
\item{object}{A drcHetVar model object}

\item{...}{Additional arguments (not used)}
}
\value{
The log-likelihood value
}
\description{
Extracts the BIC from drcHetVar model objects.
}
