% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.drcHetVar.R
\name{logLik.drcHetVar}
\alias{logLik.drcHetVar}
\title{Log-Likelihood Method for drcHetVar Objects}
\usage{
\method{logLik}{drcHetVar}(object, ...)
}
\arguments{
\item{object}{A drcHetVar model object}

\item{...}{Additional arguments (not used)}
}
\value{
The log-likelihood value
}
\description{
Extracts the log-likelihood from drcHetVar model objects.
}
