% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.drcOrdinal.R
\name{logLik.drcOrdinal}
\alias{logLik.drcOrdinal}
\title{Log-Likelihood Method for drcOrdinal Objects}
\usage{
\method{logLik}{drcOrdinal}(object, ...)
}
\arguments{
\item{object}{A drcOrdinal model object}

\item{...}{Additional arguments (not used)}
}
\value{
The log-likelihood value
}
\description{
Extracts the log-likelihood from drcOrdinal model objects.
}
