% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bmd.R
\name{print.bmd}
\alias{print.bmd}
\title{Print Method for bmd Objects}
\usage{
\method{print}{bmd}(x, ..., digits = max(3, getOption("digits") - 3))
}
\arguments{
\item{x}{A bmd model object}

\item{...}{Additional arguments (not used)}

\item{digits}{Number of significant digits to use for printing values}
}
\value{
Invisibly returns the x object
}
\description{
Prints a summary of a bmd model object.
}
