% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bmdHetVar.R
\name{print.bmdHetVar}
\alias{print.bmdHetVar}
\title{Print Method for drcHetVar Objects}
\usage{
\method{print}{bmdHetVar}(x, ..., digits = max(3, getOption("digits") - 3))
}
\arguments{
\item{x}{A drcHetVar model object}

\item{...}{Additional arguments (not used)}

\item{digits}{Number of significant digits to use for printing values}
}
\value{
Invisibly returns the x object
}
\description{
Prints a summary of a drcHetVar model object.
}
