% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.drcMMRE.R
\name{print.drcMMRE}
\alias{print.drcMMRE}
\title{Print Method for drcMMRE Objects}
\usage{
\method{print}{drcMMRE}(x, ..., digits = max(3, getOption("digits") - 3))
}
\arguments{
\item{x}{A drcMMRE model object}

\item{...}{Additional arguments (not used)}

\item{digits}{Number of significant digits to use for printing values}
}
\value{
Invisibly returns the x object
}
\description{
Prints a summary of a drcMMRE model object.
}
