% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drmHetVar.R
\name{vcov.drcHetVar}
\alias{vcov.drcHetVar}
\title{Extract Variance-Covariance Matrix from drcHetVar Objects}
\usage{
\method{vcov}{drcHetVar}(object, ...)
}
\arguments{
\item{object}{An object of class "drcHetVar"}

\item{...}{Additional arguments (currently unused)}
}
\value{
The variance-covariance matrix of the model parameters.
}
\description{
S3 method to extract the variance-covariance matrix from fitted drcHetVar objects
by inverting the Hessian of the negative log-likelihood function.
}
\seealso{
\code{\link{vcov}} for the generic function
}
