% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drmMMRE.R
\name{vcov.drcMMRE}
\alias{vcov.drcMMRE}
\title{Extract Variance-Covariance Matrix from drcMMRE Objects}
\usage{
\method{vcov}{drcMMRE}(object, ...)
}
\arguments{
\item{object}{An object of class "drcMMRE"}

\item{...}{Additional arguments (currently unused)}
}
\value{
The variance-covariance matrix of the model parameters.
}
\description{
S3 method to extract the variance-covariance matrix from fitted drcMMRE objects
by delegating to the underlying multivariate meta-analytic model.
}
\seealso{
\code{\link{vcov}} for the generic function
}
