% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootDataGenOrdinal.R
\name{bootDataGenOrdinal}
\alias{bootDataGenOrdinal}
\title{Help funtion to bmdOrdinal and bmdOrdinalMA}
\usage{
bootDataGenOrdinal(
  object,
  R = 500,
  bootType = c("nonparametric", "parametric", "model", "hierarchical")
)
}
\description{
Help funtion to bmdOrdinal and bmdOrdinalMA
}
