% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invBmd.R
\name{invBmd}
\alias{invBmd}
\title{Help funtion to bmd}
\usage{
invBmd(
  object,
  bmr,
  level = 0.9,
  slope,
  backgType = "modelBased",
  backg = NA,
  catLev = NA,
  extFactor = 10,
  def,
  useSD = useSD,
  sandwich.vcov = FALSE
)
}
\description{
Help funtion to bmd
}
