% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAV.R
\name{PAV}
\alias{PAV}
\title{Pool-adjacent-violators monotonizing}
\usage{
PAV(formula, data, type)
}
\arguments{
\item{formula}{an object of class "formula" expressing dose-response
relationship. Details of model specification are given under 'Details'}

\item{data}{data frame containing the variables in the formula}

\item{type}{character string specifying the type of data used in the model,
"continuous" or "binomial" or "Poisson"}
}
\value{
A vector containing the monotonized sequence.
}
\description{
The function monotonizes a sequence of probabilities or means based on the
pool-adjacent-violators algorithm.
}
\details{
For details on how the pool-adjacent-violators algorithm is defined see
Silvapulle and Sen (2004).

Formula should be specified as in the form number/total ~ dose for binomial
data and response ~ for continuous data.
}
\references{
Silvapulle, M. J. and Sen, P. K. (2004). Constrained statistical
inference: order, inequality, and shape constraints. New York: John Wiley &
Sons.
}
\author{
Signe M. Jensen
}
\keyword{isotonic}
\keyword{nonparametric}
\keyword{regression}
